bankwatch/requirements.txt
gunicorn
gevent
flask
plaid-python
requests[security]
redis
hiredis
stripe
